@@ -9,25 +9,23 @@
     ],
     "type": "project",
     "require": {
-        "php": "^8.0",
-        "edalzell/forma": "^1.2",
-        "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
+        "php": "^8.1",
+        "edalzell/forma": "^2.0",
+        "guzzlehttp/guzzle": "^7.2",
         "jonassiewertsen/statamic-external-link": "^1.5",
-        "laravel/framework": "^8.0",
-        "laravel/tinker": "^2.0",
+        "laravel/framework": "^10.19",
+        "laravel/tinker": "^2.8",
         "pixelfear/composer-dist-plugin": "^0.1",
-        "silentz/mailchimp": "^2.5",
-        "statamic/cms": "^3.4"
+        "silentz/mailchimp": "^4.1",
+        "statamic/cms": "^4.0"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.5",
-        "facade/ignition": "^2.3.6",
-        "fzaninotto/faker": "^1.9.1",
-        "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^10"
+        "barryvdh/laravel-debugbar": "^3.8",
+        "mockery/mockery": "^1.4.4",
+        "nunomaduro/collision": "^6.3",
+        "phpunit/phpunit": "^10",
+        "fakerphp/faker": "^1.9.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -56,7 +54,7 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "pre-update-cmd": [
@@ -74,6 +72,9 @@
             "@php artisan key:generate --ansi",
             "npm install",
             "npm run dev"
+        ],
+        "post-update-cmd": [
+            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
         ]
     }
 }
