@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\DynamicToken;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,12 +15,12 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::statamic('/search', 'search', [
-    'title' => 'Search results'
- ]);
+    'title' => 'Search results',
+]);
 
-Route::get('/!/DynamicToken/refresh', 'DynamicToken@getRefresh');
-Route::statamic('/sitemap.xml', 'sitemap/sitemap', ['layout' => null, 'content_type' => 'application/xml']); 
+Route::get('/!/DynamicToken/refresh', [DynamicToken::class, 'getRefresh']);
+Route::statamic('/sitemap.xml', 'sitemap/sitemap', ['layout' => null, 'content_type' => 'application/xml']);
 
 Route::statamic('/password', 'password', [
-    'title' => 'Passwort geschützte Seite'
- ]);
+    'title' => 'Passwort geschützte Seite',
+]);
