@@ -23,13 +23,13 @@ class Events extends CollectionTag
      */
     public function pastPerformances()
     {
-    return $this->outputDates($this->getDates('pastPerformances'));
+        return $this->outputDates($this->getDates('pastPerformances'));
     }
 
-public function future()
-{
-    return $this->outputDates($this->getDates('future'));
-}
+    public function future()
+    {
+        return $this->outputDates($this->getDates('future'));
+    }
 
     public function premieres()
     {
@@ -57,37 +57,36 @@ public function future()
         );
     }
 
-    public function count()
+    public function count(): int
     {
         return $this->getDates()->count();
     }
 
     private function getDates(string $type = 'future'): Collection
     {
-    $this->params->put('from', 'veranstaltungen');
-    $limit = $this->params->pull('limit');
+        $this->params->put('from', 'veranstaltungen');
+        $limit = $this->params->pull('limit');
 
-    $events = parent::index();
+        $events = parent::index();
 
-    if ($as = $this->params->get('as')) {
-        $events = $events[$as];
+        if ($as = $this->params->get('as')) {
+            $events = $events[$as];
+        }
+
+        $dates = $events
+            ->flatMap(fn (Entry $event) => collect($event->get('performance_dates'))
+                ->map(fn (array $date) => array_merge($event->toAugmentedArray(), $date))
+            )
+            ->sortBy(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')));
+
+        if ($type === 'pastPerformances') {
+            $dates = $dates->filter(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')) <= now());
+        } elseif ($type === 'future') {
+            $dates = $dates->filter(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')) > now());
+        }
+
+        return $dates->take($limit);
     }
-
-    $dates = $events
-        ->flatMap(fn (Entry $event) => collect($event->get('performance_dates'))
-            ->map(fn (array $date) => array_merge($event->toAugmentedArray(), $date))
-        )
-        ->sortBy(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')));
-
-    if ($type === 'pastPerformances') {
-        $dates = $dates->filter(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')) <= now());
-    } elseif ($type === 'future') {
-        $dates = $dates->filter(fn (array $event) => Carbon::parse(Arr::get($event, 'perf_date')) > now());
-    }
-
-    return $dates->take($limit);
-    }
-
 
     private function outputDates(Collection $dates): array
     {
@@ -98,10 +97,10 @@ public function future()
 
             return collect([
                 'date_groups' => $dates
-                ->groupBy(fn ($event) => Carbon::parse(Arr::get($event, 'perf_date'))->format($groupByFormat))
-                ->map(fn ($events, $date_group) => ['date_group' => $date_group, 'entries' => $events])
-                ->values()
-                ->all(),
+                    ->groupBy(fn ($event) => Carbon::parse(Arr::get($event, 'perf_date'))->format($groupByFormat))
+                    ->map(fn ($events, $date_group) => ['date_group' => $date_group, 'entries' => $events])
+                    ->values()
+                    ->all(),
             ])->all();
         }
 
