@@ -6,7 +6,7 @@ return [
 
         /*
         |--------------------------------------------------------------------------
-        | Route Prefix 
+        | Route Prefix
         |--------------------------------------------------------------------------
         |
         | The route prefix for serving HTTP based manipulated images through Glide.
@@ -35,7 +35,7 @@ return [
         |--------------------------------------------------------------------------
         |
         | The driver that will be used under the hood for image manipulation.
-        | Supported: "gd" or "imagick" (if installed on your server). 
+        | Supported: "gd" or "imagick" (if installed on your server).
         |
         */
 
