@@ -29,7 +29,7 @@ return [
     */
 
     'automatic' => env('STATAMIC_GIT_AUTOMATIC', true),
-    
+
     /*
     |--------------------------------------------------------------------------
     | Queue Connection
@@ -43,7 +43,7 @@ return [
     */
 
     'queue_connection' => env('STATAMIC_GIT_QUEUE_CONNECTION'),
-    
+
     /*
     |--------------------------------------------------------------------------
     | Dispatch Delay
@@ -147,7 +147,7 @@ return [
         // \Statamic\Events\Data\UserDeleted::class,
     ],
 
-     /*
+    /*
     |--------------------------------------------------------------------------
     | Locale
     |--------------------------------------------------------------------------
