@@ -20,7 +20,7 @@ return [
     |--------------------------------------------------------------------------
     |
     | Here you may define all of the static caching strategies for your
-    | application as well as their drivers. 
+    | application as well as their drivers.
     |
     | Supported drivers: "application", "file"
     |
@@ -36,7 +36,7 @@ return [
         'full' => [
             'driver' => 'file',
             'path' => public_path('static'),
-            'lock_hold_length' => 0
+            'lock_hold_length' => 0,
         ],
 
     ],
@@ -46,8 +46,8 @@ return [
     | Exclusions
     |--------------------------------------------------------------------------
     |
-    | Here you may define a list of URLs to be excluded from static 
-    | caching. You may want to exclude URLs containing dynamic 
+    | Here you may define a list of URLs to be excluded from static
+    | caching. You may want to exclude URLs containing dynamic
     | elements like contact forms, or shopping carts.
     |
     */
@@ -71,7 +71,6 @@ return [
     | https://docs.statamic.com/static-caching
     |
     */
-
 
     'invalidation' => [
 
