@@ -25,7 +25,7 @@ return [
                 'users' => base_path('users'),
                 'roles' => resource_path('users/roles.yaml'),
                 'groups' => resource_path('users/groups.yaml'),
-            ]
+            ],
         ],
 
         'eloquent' => [
