@@ -3,10 +3,8 @@
 /**
  * Laravel - A PHP Framework For Web Artisans
  *
- * @package  Laravel
  * @author   Taylor Otwell <taylor@laravel.com>
  */
-
 define('LARAVEL_START', microtime(true));
 
 /*
@@ -58,4 +56,4 @@ $response = $kernel->handle(
 $response->send();
 
 $kernel->terminate($request, $response);
-setlocale (LC_TIME, "de_DE.utf8");
+setlocale(LC_TIME, 'de_DE.utf8');
